create resources sub module
